BE_Add_Course_Module

+++ CourseManagementBackend/src/main/java/com/app/dao/SubjectDao.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.app.entity.Subject;
 
-public interface SubjectDao extends JpaRepository<Subject, Long> {
+public interface SubjectDao extends JpaRepository<Subject, Integer> {
 
 }
